@@ -1,2 +1,2 @@
-export 'src/public/app_store.dart';
+export 'src/public/flutter_launch_store.dart';
 export 'src/public/model/app_store_model.dart';
